+++ web/src/services/api/index.ts
@@ -1,0 +1,2 @@
+export { useProjectFetcher } from "./projectApi";
+export { useTeamsFetcher, useCreateTeamFetcher } from "./teamsApi";
